Working example of system/templates/admin overloading
This adds a system_blocks_item.tpl with a simple change of
all the class tooltip references changed to xxtooltip.

With change the action icons are restored on the blocks
display on system/admin.php?fct=blocksadmin

This example shows the way to overload the existing template
with bootstrap compatible ones.